Two main sets of changes: Full key binding support (custom keyboard layouts) and better organization of shared materials.
First is key binding. there is a new file called "keymap.txt" that allows players to bind any of the ingame commands to any of the keys in the XNA-defined Keys enum, as well as a few special cases (control, alt, shift).
There is no support for macros (eg. you can't have a key fire then switch weapon) as each key can only do one action for now.
If you mess up your key bindings you can just delete the file to get the original back.
The key binding file supports comments with # and any spacing.


Second is some code structure changes. Most notably, I've moved a lot of the shared classes / constants to a new project called "InfiniminerShared". This makes it easier to find a lot of the definitions that are used across multiple classes and/or across the client and server.
This is also in preperation of a proper client-server system - that was actually the initial intention of the Buttons enum, but I then pursued custom key layouts.

Signed-off-by: Zach Barth <zachbarth@gmail.com>